TrelloList refactoring, rename helpers folder and other changes
